requirements-dev.txt
codecov
flake8
mypy
nbval
numpydoc
pylint
pytest
pytest-cov
sphinx
sphinx-gallery
twine
yapf
